Make updates suggested by new version of Xcode https://bugs.webkit.org/show_bug.cgi?id=136603
Reviewed by Mark Rowe.

Source/JavaScriptCore:

* Configurations/Base.xcconfig: Added CLANG_WARN_UNREACHABLE_CODE, COMBINE_HIDPI_IMAGES,
and ENABLE_STRICT_OBJC_MSGSEND as suggested by Xcode upgrade check.

* JavaScriptCore.xcodeproj/project.pbxproj: Update LastUpgradeCheck.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::jumpSlowForUnwantedArrayMode): Compile out unreachable code
for clang, since it understands the code is unreachable.
* runtime/JSArray.cpp:
(JSC::JSArray::fillArgList): Ditto.
(JSC::JSArray::copyToArguments): Ditto.

Source/ThirdParty:

* gtest/xcode/Config/General.xcconfig: Moved CLANG_WARN_BOOL_CONVERSION,
CLANG_WARN_ENUM_CONVERSION, CLANG_WARN_INT_CONVERSION, and COMBINE_HIDPI_IMAGES
here from project file.  Added CLANG_WARN_UNREACHABLE_CODE, ENABLE_STRICT_OBJC_MSGSEND,
and GCC_WARN_64_TO_32_BIT_CONVERSION as suggested by Xcode upgrade check.

* gtest/xcode/gtest.xcodeproj/project.pbxproj: Updated LastUpgradeCheck and removed
things that are redundant with the xcconfig file above.

Source/ThirdParty/ANGLE:

* ANGLE.xcodeproj/project.pbxproj: Updated LastUpgradeCheck.

* Configurations/ANGLE.xcconfig: Added CLANG_WARN_BOOL_CONVERSION,
CLANG_WARN_ENUM_CONVERSION, CLANG_WARN_INT_CONVERSION, COMBINE_HIDPI_IMAGES,
and ENABLE_STRICT_OBJC_MSGSEND as suggested by Xcode upgrade check.

Source/WebCore:

* Configurations/Base.xcconfig: Added COMBINE_HIDPI_IMAGES and
ENABLE_STRICT_OBJC_MSGSEND as suggested by Xcode upgrade check.

* WebCore.xcodeproj/project.pbxproj: Let Xcode remove an orphaned item from the file,
and also updated LastUpgradeCheck.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::updateTextTrackDisplay): Use #if/#else/#endif instead of
using #if/return/#endif and leaving code unreachable.
(WebCore::HTMLMediaElement::configureMediaControls): Ditto.
(WebCore::HTMLMediaElement::configureTextTrackDisplay): Ditto.
* html/canvas/WebGLDrawBuffers.cpp:
(WebCore::WebGLDrawBuffers::supported): Ditto.

Source/WebInspectorUI:

* Configurations/Base.xcconfig: Added CLANG_WARN_BOOL_CONVERSION, CLANG_WARN_CONSTANT_CONVERSION,
CLANG_WARN_EMPTY_BODY, CLANG_WARN_ENUM_CONVERSION, CLANG_WARN_INT_CONVERSION,
CLANG_WARN_UNREACHABLE_CODE, CLANG_WARN__DUPLICATE_METHOD_MATCH, ENABLE_STRICT_OBJC_MSGSEND,
GCC_WARN_UNDECLARED_SELECTOR, and GCC_WARN_UNUSED_FUNCTION as suggested by Xcode update check.
Also removed duplicate GCC_WARN_UNUSED_VARIABLE line.

* WebInspectorUI.xcodeproj/project.pbxproj: Updated LastUpgradeCheck.

Source/WebKit:

* WebKit.xcodeproj/project.pbxproj: Updated LastUpgradeCheck.

Source/WebKit/mac:

* Configurations/Base.xcconfig: Added CLANG_WARN_UNREACHABLE_CODE, COMBINE_HIDPI_IMAGES,
and ENABLE_STRICT_OBJC_MSGSEND as suggested by Xcode upgrade check.

* History/WebURLsWithTitles.m:
(+[WebURLsWithTitles writeURLs:andTitles:toPasteboard:]): Use NSUInteger instead of
unsigned for correctness, and to quiet the 32-to-64-bit compiler warning (which is
not turned on because it gives us too many false positives).
* Misc/WebElementDictionary.mm:
(-[WebElementDictionary objectForKey:]): Ditto.
* Misc/WebNSDataExtras.m:
(-[NSData _webkit_guessedMIMETypeForXML]): Ditto.
(-[NSData _webkit_guessedMIMEType]): Ditto.
(-[NSData _webkit_parseRFC822HeaderFields]): Ditto.
(-[NSData _web_locationAfterFirstBlankLine]): Ditto.

* Misc/WebNSURLExtras.h: Removed unused _web_URLWithLowercasedScheme, _web_hostData,
_webkit_URLByRemovingFragment, _webkit_URLByRemovingResourceSpecifier, _webkit_isFTPDirectoryURL,
_webkit_shouldLoadAsEmptyDocument, _web_hostNameNeedsDecodingWithRange:,
_web_hostNameNeedsEncodingWithRange:, _web_decodeHostNameWithRange:, _web_encodeHostNameWithRange:,
and _webkit_URLFragment methods.

* Misc/WebNSURLExtras.mm:
(-[NSURL _web_URLByTruncatingOneCharacterBeforeComponent:]): Deleted.
(-[NSURL _webkit_URLByRemovingFragment]): Deleted.
(-[NSURL _webkit_URLByRemovingResourceSpecifier]): Deleted.
(-[NSURL _webkit_isFTPDirectoryURL]): Deleted.
(-[NSURL _webkit_shouldLoadAsEmptyDocument]): Deleted.
(-[NSURL _web_URLWithLowercasedScheme]): Deleted.
(-[NSString _webkit_isFTPDirectoryURL]): Deleted.
(-[NSString _web_hostNameNeedsDecodingWithRange:]): Deleted.
(-[NSString _web_hostNameNeedsEncodingWithRange:]): Deleted.
(-[NSString _web_decodeHostNameWithRange:]): Deleted.
(-[NSString _web_encodeHostNameWithRange:]): Deleted.
(-[NSString _webkit_URLFragment]): Deleted.

* Plugins/WebPluginDatabase.mm:
(-[WebPluginDatabase removePluginInstanceViewsFor:]): Use a modern for loop
to iterate an array instead of a loop using the type "unsigned int".

* WebCoreSupport/WebOpenPanelResultListener.mm:
(-[WebOpenPanelResultListener chooseFilenames:]): Use NSUInteger instead of int.

* WebView/WebDelegateImplementationCaching.mm:
(CallDelegate): Use wtfCallIMP in one overload that was instead just calling the
directly without a proper type for the function pointer. This might have been causing
an actual problem on 64-bit systems, and it was different from all the other CallDelegate
functions that were already doing this correctly.

* WebView/WebTextCompletionController.mm:
(-[WebTextCompletionController _placePopupWindow:]): Use NSUInteger instead of int.

Source/WebKit2:

* Configurations/Base.xcconfig: Added CLANG_WARN_UNREACHABLE_CODE, COMBINE_HIDPI_IMAGES,
and ENABLE_STRICT_OBJC_MSGSEND as suggested by the Xcode upgrade check.

* PluginProcess/mac/PluginProcessMac.mm:
(WebKit::initializeCocoaOverrides): Changed NSConcreteTask_launch to use the same technique
as the other functions in this file, doing the type casting right at the call to
method_setImplementation so the global has a suitable type for making a function call,
rather than relying on the abilty to call through a type without a specific argument list.

* UIProcess/Plugins/PluginInfoStore.cpp:
(WebKit::pathExtension): Use size_t for the result of String::reverseFind rather than
converting it to an int. Also don't rely on the fact that WTF's notFound becomes -1 when
cast from size_t to int.

* WebKit2.xcodeproj/project.pbxproj: Let Xcode delete some orphaned items, and updated
LastUpgradeCheck.

Source/WTF:

* Configurations/Base.xcconfig: Added CLANG_WARN_UNREACHABLE_CODE,
COMBINE_HIDPI_IMAGES, and ENABLE_STRICT_OBJC_MSGSEND as suggested by
Xcode upgrade check.

* WTF.xcodeproj/project.pbxproj: Updated LastUpgradeCheck.

Canonical link: https://commits.webkit.org/154424@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173364 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/mac/History/WebURLsWithTitles.m
@@ -51,7 +51,7 @@
 {
     NSMutableArray *URLStrings;
     NSMutableArray *titlesOrEmptyStrings;
-    unsigned index, count;
+    NSUInteger index, count;
 
     count = [URLs count];
     if (count == 0) {
